+++ src/ja/comicmedu/src/eu/kanade/tachiyomi/extension/ja/comicmedu/ComicMeDu.kt
@@ -1,17 +1,11 @@
 package eu.kanade.tachiyomi.extension.ja.comicmedu
 
-import eu.kanade.tachiyomi.multisrc.comiciviewer.ComiciViewerAlt
+import eu.kanade.tachiyomi.multisrc.comiciviewer.ComiciViewer
 import eu.kanade.tachiyomi.network.GET
+import keiyoushi.annotation.Source
 import okhttp3.Request
 
-class ComicMeDu :
-    ComiciViewerAlt(
-        "G-Comi",
-        "https://g-comi.jp",
-        "ja",
-        "https://g-comi.jp/api",
-    ) {
-    override val id = 7310112963091407823
-
+@Source
+abstract class ComicMeDu : ComiciViewer() {
     override fun latestUpdatesRequest(page: Int): Request = GET("$baseUrl/category/manga/$page", headers)
 }
