+++ src/ja/comicpash/src/eu/kanade/tachiyomi/extension/ja/comicpash/ComicPash.kt
@@ -1,11 +1,7 @@
 package eu.kanade.tachiyomi.extension.ja.comicpash
 
-import eu.kanade.tachiyomi.multisrc.comiciviewer.ComiciViewerAlt
+import eu.kanade.tachiyomi.multisrc.comiciviewer.ComiciViewer
+import keiyoushi.annotation.Source
 
-class ComicPash :
-    ComiciViewerAlt(
-        "Comic Pash",
-        "https://comicpash.jp",
-        "ja",
-        "https://comicpash.jp/api",
-    )
+@Source
+abstract class ComicPash : ComiciViewer()
