+++ src/ja/comicride/build.gradle.kts
@@ -4,10 +4,13 @@ plugins {
 
 keiyoushi {
     name = "Comic Ride"
-    className = "ComicRide"
     versionCode = 0
     contentWarning = ContentWarning.SAFE
     libVersion = "1.4"
     theme = "comiciviewer"
-    baseUrl = "https://comicride.jp"
+
+    source {
+        lang = "ja"
+        baseUrl = "https://comicride.jp"
+    }
 }
