+++ src/ja/hanayume/build.gradle.kts
@@ -4,10 +4,13 @@ plugins {
 
 keiyoushi {
     name = "Hana To Yume+"
-    className = "Hanayume"
     versionCode = 0
     contentWarning = ContentWarning.SAFE
     libVersion = "1.4"
     theme = "comiciviewer"
-    baseUrl = "https://hanayume.com"
+
+    source {
+        lang = "ja"
+        baseUrl = "https://hanayume.com"
+    }
 }
