+++ src/ja/herosweb/build.gradle.kts
@@ -4,10 +4,13 @@ plugins {
 
 keiyoushi {
     name = "HERO'S Web"
-    className = "HerosWeb"
     versionCode = 0
     contentWarning = ContentWarning.SAFE
     libVersion = "1.4"
     theme = "comiciviewer"
-    baseUrl = "https://heros-web.com"
+
+    source {
+        lang = "ja"
+        baseUrl = "https://heros-web.com"
+    }
 }
