+++ src/ja/jnbooks/src/eu/kanade/tachiyomi/extension/ja/jnbooks/JNBooks.kt
@@ -1,18 +1,14 @@
 package eu.kanade.tachiyomi.extension.ja.jnbooks
 
-import eu.kanade.tachiyomi.multisrc.comiciviewer.ComiciViewerAlt
+import eu.kanade.tachiyomi.multisrc.comiciviewer.ComiciViewer
 import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.source.model.MangasPage
+import keiyoushi.annotation.Source
 import okhttp3.Request
 import okhttp3.Response
 
-class JNBooks :
-    ComiciViewerAlt(
-        "J-N Books",
-        "https://comic.j-nbooks.jp",
-        "ja",
-        "https://comic.j-nbooks.jp/api",
-    ) {
+@Source
+abstract class JNBooks : ComiciViewer() {
     override val supportsLatest = false
 
     override fun popularMangaRequest(page: Int): Request = GET("$baseUrl/category/manga/$page", headers)
