+++ src/ja/kimicomi/build.gradle.kts
@@ -4,10 +4,13 @@ plugins {
 
 keiyoushi {
     name = "KimiComi"
-    className = "KimiComi"
     versionCode = 1
     contentWarning = ContentWarning.SAFE
     libVersion = "1.4"
     theme = "comiciviewer"
-    baseUrl = "https://kimicomi.com"
+
+    source {
+        lang = "ja"
+        baseUrl = "https://kimicomi.com"
+    }
 }
