+++ src/ja/magkan/build.gradle.kts
@@ -4,10 +4,13 @@ plugins {
 
 keiyoushi {
     name = "MagKan"
-    className = "MagKan"
     versionCode = 1
     contentWarning = ContentWarning.SAFE
     libVersion = "1.4"
     theme = "comiciviewer"
-    baseUrl = "https://kansai.mag-garden.co.jp"
+
+    source {
+        lang = "ja"
+        baseUrl = "https://kansai.mag-garden.co.jp"
+    }
 }
