+++ src/ja/magkan/src/eu/kanade/tachiyomi/extension/ja/magkan/MagKan.kt
@@ -1,18 +1,14 @@
 package eu.kanade.tachiyomi.extension.ja.magkan
 
-import eu.kanade.tachiyomi.multisrc.comiciviewer.ComiciViewerAlt
+import eu.kanade.tachiyomi.multisrc.comiciviewer.ComiciViewer
 import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.source.model.MangasPage
+import keiyoushi.annotation.Source
 import okhttp3.Request
 import okhttp3.Response
 
-class MagKan :
-    ComiciViewerAlt(
-        "MagKan",
-        "https://kansai.mag-garden.co.jp",
-        "ja",
-        "https://kansai.mag-garden.co.jp/api",
-    ) {
+@Source
+abstract class MagKan : ComiciViewer() {
     override val supportsLatest = false
 
     override fun popularMangaRequest(page: Int): Request = GET("$baseUrl/category/manga/$page", headers)
