+++ src/ja/mangabang/build.gradle.kts
@@ -4,10 +4,13 @@ plugins {
 
 keiyoushi {
     name = "MangaBang Comics"
-    className = "MangaBang"
     versionCode = 1
     contentWarning = ContentWarning.SAFE
     libVersion = "1.4"
     theme = "comiciviewer"
-    baseUrl = "https://comics.manga-bang.com"
+
+    source {
+        lang = "ja"
+        baseUrl = "https://comics.manga-bang.com"
+    }
 }
