+++ src/ja/mangabang/src/eu/kanade/tachiyomi/extension/ja/mangabang/MangaBang.kt
@@ -1,16 +1,12 @@
 package eu.kanade.tachiyomi.extension.ja.mangabang
 
-import eu.kanade.tachiyomi.multisrc.comiciviewer.ComiciViewerAlt
+import eu.kanade.tachiyomi.multisrc.comiciviewer.ComiciViewer
 import eu.kanade.tachiyomi.network.GET
+import keiyoushi.annotation.Source
 import okhttp3.Request
 
-class MangaBang :
-    ComiciViewerAlt(
-        "MangaBang Comics",
-        "https://comics.manga-bang.com",
-        "ja",
-        "https://comics.manga-bang.com/api",
-    ) {
+@Source
+abstract class MangaBang : ComiciViewer() {
     override val rankingFromNextJs = false
 
     override fun latestUpdatesRequest(page: Int): Request = GET("$baseUrl/category/manga/$page", headers)
