+++ src/ja/mangacross/build.gradle.kts
@@ -4,10 +4,14 @@ plugins {
 
 keiyoushi {
     name = "Champion Cross"
-    className = "MangaCross"
     versionCode = 5
     contentWarning = ContentWarning.SAFE
     libVersion = "1.4"
     theme = "comiciviewer"
-    baseUrl = "https://championcross.jp"
+
+    source {
+        lang = "ja"
+        baseUrl = "https://championcross.jp"
+        versionId = 2
+    }
 }
