+++ src/ja/rimacomiplus/build.gradle.kts
@@ -4,10 +4,13 @@ plugins {
 
 keiyoushi {
     name = "RimacomiPlus"
-    className = "RimacomiPlus"
     versionCode = 0
     contentWarning = ContentWarning.SAFE
     libVersion = "1.4"
     theme = "comiciviewer"
-    baseUrl = "https://rimacomiplus.jp"
+
+    source {
+        lang = "ja"
+        baseUrl = "https://rimacomiplus.jp"
+    }
 }
