+++ src/ja/takecomic/build.gradle.kts
@@ -4,10 +4,13 @@ plugins {
 
 keiyoushi {
     name = "TakeComic"
-    className = "TakeComic"
     versionCode = 1
     contentWarning = ContentWarning.SAFE
     libVersion = "1.4"
     theme = "comiciviewer"
-    baseUrl = "https://takecomic.jp"
+
+    source {
+        lang = "ja"
+        baseUrl = "https://takecomic.jp"
+    }
 }
