+++ src/ja/younganimal/build.gradle.kts
@@ -4,10 +4,13 @@ plugins {
 
 keiyoushi {
     name = "Young Animal"
-    className = "YoungAnimal"
     versionCode = 0
-    contentWarning = ContentWarning.SAFE
+    contentWarning = ContentWarning.MIXED
     libVersion = "1.4"
     theme = "comiciviewer"
-    baseUrl = "https://younganimal.com"
+
+    source {
+        lang = "ja"
+        baseUrl = "https://younganimal.com"
+    }
 }
