+++ src/ja/youngchampion/build.gradle.kts
@@ -4,10 +4,13 @@ plugins {
 
 keiyoushi {
     name = "Young Champion"
-    className = "YoungChampion"
     versionCode = 0
-    contentWarning = ContentWarning.NSFW
+    contentWarning = ContentWarning.MIXED
     libVersion = "1.4"
     theme = "comiciviewer"
-    baseUrl = "https://youngchampion.jp"
+
+    source {
+        lang = "ja"
+        baseUrl = "https://youngchampion.jp"
+    }
 }
